requirements.txt
Flask>=2.0
jsonschema>=4.0
PyYAML>=6.0